+++ src/entities/post/model/post.types.ts
