--- src/entities/post/ui/index.ts
@@ -1,2 +1,0 @@
-export { PostsTable } from "./posts-table";
-export type { PostsTableProps } from "./posts-table";
