--- src/entities/user/model/types.ts
@@ -1,1 +1,0 @@
-export type User = { id: number; username: string; image?: string };
