+++ src/entities/user/model/user.types.ts
@@ -1,0 +1,1 @@
+export type User = { id: number; username: string; image?: string; fullname?: string };
