+++ src/features/edit-post/index.ts
@@ -1,0 +1,3 @@
+export { usePostEditor } from "./model/edit-post.hook";
+export type { CreatePostParams } from "./api/create-post.api";
+export type { UpdatePostPayload } from "./api/update-post.api";
