+++ src/features/edit-post/model/edit-post.hook.ts
@@ -2,9 +2,9 @@ import { useState } from "react";
 
 import { usePost } from "@entities/post";
 
-import { createPost, type CreatePostParams } from "../api/create-post";
-import { deletePostRequest } from "../api/delete-post";
-import { updatePost as updatePostRequest, type UpdatePostPayload } from "../api/update-post";
+import { createPost, type CreatePostParams } from "../api/create-post.api";
+import { deletePostRequest } from "../api/delete-post.api";
+import { updatePost as updatePostRequest, type UpdatePostPayload } from "../api/update-post.api";
 
 export function usePostEditor() {
   const { posts, appendPost, changePost, removePost } = usePost();
