--- src/features/post-editor/index.ts
@@ -1,3 +1,0 @@
-export { usePostEditor } from "./model/use-post-editor";
-export type { CreatePostParams } from "./api/create-post";
-export type { UpdatePostPayload } from "./api/update-post";
